refactor and fix screenshot save
