+++ openhands/runtime/browser/utils.py
@@ -1,7 +1,11 @@
 import base64
 import datetime
 import os
+import base64
+import io
 from pathlib import Path
+from PIL import Image
+import numpy as np
 
 from openhands.core.exceptions import BrowserUnavailableException
 from openhands.core.schema import ActionType
@@ -10,6 +14,32 @@ from openhands.events.observation import BrowserOutputObservation
 from openhands.runtime.browser.browser_env import BrowserEnv
 from openhands.utils.async_utils import call_sync_from_async
 
+def image_to_png_base64_url(
+    image: np.ndarray | Image.Image, add_data_prefix: bool = False
+) -> str:
+    """Convert a numpy array to a base64 encoded png image url."""
+    if isinstance(image, np.ndarray):
+        image = Image.fromarray(image)
+    if image.mode in ('RGBA', 'LA'):
+        image = image.convert('RGB')
+    buffered = io.BytesIO()
+    image.save(buffered, format='PNG')
+
+    image_base64 = base64.b64encode(buffered.getvalue()).decode()
+    return (
+        f'data:image/png;base64,{image_base64}'
+        if add_data_prefix
+        else f'{image_base64}'
+    )
+
+def png_base64_url_to_image(png_base64_url: str) -> Image.Image:
+    """Convert a base64 encoded png image url to a PIL Image."""
+    splited = png_base64_url.split(',')
+    if len(splited) == 2:
+        base64_data = splited[1]
+    else:
+        base64_data = png_base64_url
+    return Image.open(io.BytesIO(base64.b64decode(base64_data)))
 
 async def browse(
     action: BrowseURLAction | BrowseInteractiveAction,
@@ -50,9 +80,8 @@ async def browse(
             screenshot_path = str(screenshots_dir / screenshot_filename)
 
             # Decode and save the screenshot
-            screenshot_data = base64.b64decode(obs.get('screenshot'))
-            with open(screenshot_path, 'wb') as f:
-                f.write(screenshot_data)
+            image = png_base64_url_to_image(obs.get('screenshot'))
+            image.save(screenshot_path)
 
         return BrowserOutputObservation(
             content=obs['text_content'],  # text content of the page
